requirements.txt
# Main dependencies
fastapi==0.104.1
uvicorn==0.24.0.post1
pydantic==2.4.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic-settings==2.0.3
alembic

# Authentication & Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# Async functionality
httpx==0.25.1
asyncpg==0.28.0
aiosqlite

# Task queue
celery==5.3.4
redis==5.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
locust

# Utilities
python-dotenv==1.0.0
email-validator==2.1.0
minio==7.1.16
cryptography==41.0.7
pyclamd==0.4.0
aiobotocore==2.7.0

# Rate limiting
fastapi-limiter